api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.2
